2_prompt_chatbots/requirements.txt
langchain
langchain-community
langchain_groq
python-dotenv
streamlit